+++ label_sorter/__init__.py
@@ -1,1 +1,1 @@
-from .core import LabelSorter
+from .sorter import LabelSorter
