+++ label_sorter/platforms/ecommerce/shopify.py
@@ -1,5 +1,5 @@
 import re
-from .base_label import BaseLabel
+from ..base_label import BaseLabel
 
 class ShopifyLabel(BaseLabel):
     def __init__(self, page_text, page_table,page_num):
